requirements.txt
kafka-python==2.0.2
pyspark==3.5.0
pandas==2.1.0
pyarrow==14.0.0
boto3==1.34.0
psycopg2-binary==2.9.9
prometheus-client==0.18.0
numpy==1.26.0
